controller: Overhaul for impulses and forces
'Simulate' with a fixed time step. Handle input events that happened during
that fixed time step. This works if the game loop proceeds faster than the
simulation time step but could break down if the frame rate is below 60fps.
Ideally events would have timestamps to correlate them with the physics time
step to which they belong.

A desired velocity in x,z is calculated based on input events and configured
walk / run speed. If jump was just pressed, the desired y velocity is set to
the jump speed. The difference between the current velocity in x,z and the
desired velocity is calculated (to excluding velocity due to falling, which is
the responsibility of the physics engine, not input). The impulse (change in
momentum) is calculated by multiplying by the mass of the body. The force (rate
of change of momentum) is calculated by dividing the impulse by the time step
(as the force is only active for a single time step.) If we were jumping, we
apply gravity to the y-component based on the previous y-velocity, else we just
use the desired velocity as-is and calculate the translation by multiplying by
the time step. Then we reset the input state because all inputs were handled
for a single time step of the simulation.